requirements.txt
pyrofork==2.3.68
fastapi==0.128.0
uvicorn==0.38.0
asyncpg==0.31.0
python-dotenv==1.2.1
redis==7.1.0
TgCrypto-pyrofork==1.2.8
passlib[argon2]
argon2-cffi==25.1.0
python-multipart==0.0.20
itsdangerous==2.2.0
Jinja2==3.1.6